Moved the data access code into their own modules.
